requirements.txt
numpy==1.19.5
pandas==1.4.1
tensorflow==2.6.0
keras==2.6.0
google-cloud-storage==2.2.1
google-cloud-bigquery==3.0.1
google-auth==1.33.1
ipykernel==6.0.3
jupyterlab==3.0.16
pyarrow==7.0.0
pytextdist==0.1.6
gensim==3.8.3
matplotlib==3.4.3
scikit-learn==1.0.2
nltk==3.6.2
annoy==1.17.0
seaborn==0.11.2
textdistance==4.2.2
modin[all]==0.14.0
papermill==2.3.4
Unidecode==1.3.4
click==8.0.4